// src/lib/form-schema.ts
import z from "zod";

export type FormType = "Login" | "Sign Up";

export const loginSchema = z.object({
  email: z.email("Invalid email address").min(1, "Email is required"),
  password: z.string().min(1, "Password is required"),
});

export const signUpSchema = z.object({
  name: z.string().min(1, "Full name is required"),
  email: z.email("Invalid email address").min(1, "Email is required"),
  password: z.string().min(8, "Password must be at least 8 characters long"),
});

export type PlanType = "Basic" | "Advanced" | "Business";